# Refactor Migration for better performance