--- a/x/metadata/keeper/migrations.go
+++ b/x/metadata/keeper/migrations.go
@@ -1,7 +1,6 @@
 package keeper
 
 import (
-	"bytes"
 	"fmt"
 
 	"github.com/cosmos/cosmos-sdk/store/prefix"
@@ -32,8 +31,9 @@ func (m *Migrator) Migrate1to2(ctx sdk.Context) error {
 // Migrate2to3 migrates from version 2 to 3 to fix the indexes.
 func (m *Migrator) Migrate2to3(ctx sdk.Context) error {
 	ctx.Logger().Info("Migrating Metadata Module from Version 2 to 3")
-	var goodIndexes *indexLookup
+	var goodIndexes keyLookup
 	var sessionsToDelete []types.MetadataAddress
+	var sessionsToKeep keyLookup
 	steps := []struct {
 		name   string
 		runner func() error
@@ -47,22 +47,30 @@ func (m *Migrator) Migrate2to3(ctx sdk.Context) error {
 			},
 		},
 		{
-			name:   "Reindex scopes",
+			name:   "Reindexing scopes",
 			runner: func() error { return reindexScopes(ctx, m.keeper, goodIndexes) },
 		},
 		{
-			name:   "Reindex scope specs",
+			name:   "Reindexing scope specs",
 			runner: func() error { return reindexScopeSpecs(ctx, m.keeper, goodIndexes) },
 		},
 		{
-			name:   "Reindex contract specs",
+			name:   "Reindexing contract specs",
 			runner: func() error { return reindexContractSpecs(ctx, m.keeper, goodIndexes) },
+		},
+		{
+			name: "Identifying sessions to keep",
+			runner: func() error {
+				var err error
+				sessionsToKeep, err = identifySessionsWithRecords(ctx, m.keeper)
+				return err
+			},
 		},
 		{
 			name: "Finding empty sessions",
 			runner: func() error {
 				var err error
-				sessionsToDelete, err = getEmptySessions(ctx, m.keeper)
+				sessionsToDelete, err = identifySessionsToDelete(ctx, m.keeper, sessionsToKeep)
 				return err
 			},
 		},
@@ -85,66 +93,38 @@ func (m *Migrator) Migrate2to3(ctx sdk.Context) error {
 	return nil
 }
 
-// indexLookup is a struct holding index byte slices in a way that makes them easier to find.
-type indexLookup struct {
-	// entries is a map of the first keyLen bytes to all indexes that start with those bytes.
-	entries map[string][][]byte
-	// pkLen is the number of bytes from index to use as the key for entries.
-	pkLen int
+// keyLookup is a map used to identify known keys.
+type keyLookup map[string]struct{}
+
+// newKeyLookup creates a new empty keyLookup.
+func newKeyLookup() keyLookup {
+	return map[string]struct{}{}
 }
 
-// newIndexLookup creates a new empty IndexLookup.
-func newIndexLookup() *indexLookup {
-	return &indexLookup{
-		entries: map[string][][]byte{},
-		// All the keys we have so far have length 20.
-		// Add 1 for the type byte, and 1 for length byte. Then also use the first 2 of metadata address.
-		// 20 + 1 + 1 + 2 = 24.
-		pkLen: 24,
-	}
-}
-
-// getPK creates a primary key string for the provided index key.
-func (i indexLookup) getPK(indexKey []byte) string {
-	if len(indexKey) <= i.pkLen {
-		return string(indexKey)
-	}
-	return string(indexKey[:i.pkLen])
-}
-
-// add records the given index keys in this indexLookup.
-func (i *indexLookup) add(indexKeys ...[]byte) {
-	for _, indexKey := range indexKeys {
-		pk := i.getPK(indexKey)
-		i.entries[pk] = append(i.entries[pk], indexKey)
-	}
-}
-
-// has returns true if the provided indexKey is known to this indexLookup.
-func (i indexLookup) has(indexKey []byte) bool {
-	if len(indexKey) == 0 {
-		return false
-	}
-	l, ok := i.entries[i.getPK(indexKey)]
-	if !ok {
-		return false
-	}
-	for _, k := range l {
-		if bytes.Equal(indexKey, k) {
-			return true
+// add records the given index keys in this keyLookup.
+func (m *keyLookup) add(indexKeys ...[]byte) {
+	for _, key := range indexKeys {
+		k := string(key)
+		if _, has := (*m)[k]; !has {
+			(*m)[k] = struct{}{}
 		}
 	}
-	return false
 }
 
-// deleteBadIndexes deletes all the bad indexes on Metadata entries, and gets the metadata addresses of things de-indexed.
+// has returns true if the provided indexKey is known to this keyLookup.
+func (m keyLookup) has(indexKey []byte) bool {
+	_, rv := m[string(indexKey)]
+	return rv
+}
+
+// deleteBadIndexes deletes all the bad indexes on Metadata entries, and gets the metadata addresses of things properly indexed.
 // This is a function for a migration, not intended for outside use.
-func deleteBadIndexes(ctx sdk.Context, mdKeeper Keeper) (*indexLookup, error) {
+func deleteBadIndexes(ctx sdk.Context, mdKeeper Keeper) (keyLookup, error) {
 	prefixes := [][]byte{
 		types.AddressScopeCacheKeyPrefix, types.ValueOwnerScopeCacheKeyPrefix,
 		types.AddressScopeSpecCacheKeyPrefix, types.AddressContractSpecCacheKeyPrefix,
 	}
-	good := newIndexLookup()
+	good := newKeyLookup()
 	store := ctx.KVStore(mdKeeper.storeKey)
 	for _, pre := range prefixes {
 		newGood, err := cleanStore(ctx, store, pre)
@@ -153,6 +133,7 @@ func deleteBadIndexes(ctx sdk.Context, mdKeeper Keeper) (*indexLookup, error) {
 		}
 		good.add(newGood...)
 	}
+	ctx.Logger().Info("Done identifying good indexes and deleting bad ones.")
 	return good, nil
 }
 
@@ -174,7 +155,10 @@ func cleanStore(ctx sdk.Context, baseStore sdk.KVStore, pre []byte) (good [][]by
 	defer func() {
 		err = iter.Close()
 	}()
+	i := 0
+	ctx.Logger().Info(fmt.Sprintf("Prefix %X: Identifying good entries and deleting bad ones.", pre))
 	for ; iter.Valid(); iter.Next() {
+		i++
 		key := iter.Key()
 		accAddrLen := int(key[0])
 		// good key length = 1 length byte + the length of the account address + 17 metadata address bytes.
@@ -184,67 +168,141 @@ func cleanStore(ctx sdk.Context, baseStore sdk.KVStore, pre []byte) (good [][]by
 			store.Delete(key)
 			deletedCount++
 		}
+		if i%10000 == 0 {
+			ctx.Logger().Info(fmt.Sprintf("Prefix %X: Deleted %d entries, found %d good entries so far.", pre, deletedCount, len(good)))
+		}
 	}
-	ctx.Logger().Info(fmt.Sprintf("Prefix %X: Deleted %d entries, found %d good entries.", pre, deletedCount, len(good)))
+	ctx.Logger().Info(fmt.Sprintf("Prefix %X: Done. Deleted %d entries, found %d good entries.", pre, deletedCount, len(good)))
 	return good, nil
 }
 
 // reindexScopes creates all missing scope indexes.
 // This is a function for a migration, not intended for outside use.
-func reindexScopes(ctx sdk.Context, mdKeeper Keeper, lookup *indexLookup) error {
+func reindexScopes(ctx sdk.Context, mdKeeper Keeper, lookup keyLookup) error {
 	store := ctx.KVStore(mdKeeper.storeKey)
-	return mdKeeper.IterateScopes(ctx, func(scope types.Scope) (stop bool) {
+	i := 0
+	ri := 0
+	rv := mdKeeper.IterateScopes(ctx, func(scope types.Scope) (stop bool) {
+		i++
+		didri := false
 		for _, key := range getScopeIndexValues(&scope).IndexKeys() {
 			if (key[0] == types.AddressScopeCacheKeyPrefix[0] || key[0] == types.ValueOwnerScopeCacheKeyPrefix[0]) && !lookup.has(key) {
 				store.Set(key, []byte{0x01})
+				didri = true
 			}
+		}
+		if didri {
+			ri++
+		}
+		if i%10000 == 0 {
+			ctx.Logger().Info(fmt.Sprintf("Checked %d scopes and reindexed %d of them.", i, ri))
 		}
 		return false
 	})
+	ctx.Logger().Info(fmt.Sprintf("Done reindexing %d scopes. All %d are now indexed.", ri, i))
+	return rv
 }
 
 // reindexScopeSpecs creates all missing scope specification indexes.
 // This is a function for a migration, not intended for outside use.
-func reindexScopeSpecs(ctx sdk.Context, mdKeeper Keeper, lookup *indexLookup) error {
+func reindexScopeSpecs(ctx sdk.Context, mdKeeper Keeper, lookup keyLookup) error {
 	store := ctx.KVStore(mdKeeper.storeKey)
-	return mdKeeper.IterateScopeSpecs(ctx, func(scopeSpec types.ScopeSpecification) (stop bool) {
+	i := 0
+	ri := 0
+	rv := mdKeeper.IterateScopeSpecs(ctx, func(scopeSpec types.ScopeSpecification) (stop bool) {
+		i++
+		didri := false
 		for _, key := range getScopeSpecIndexValues(&scopeSpec).IndexKeys() {
 			if key[0] == types.AddressScopeSpecCacheKeyPrefix[0] && !lookup.has(key) {
 				store.Set(key, []byte{0x01})
+				didri = true
 			}
+		}
+		if didri {
+			ri++
+		}
+		if i%10000 == 0 {
+			ctx.Logger().Info(fmt.Sprintf("Checked %d scope specs and reindexed %d of them.", i, ri))
 		}
 		return false
 	})
+	ctx.Logger().Info(fmt.Sprintf("Done reindexing %d scope specs. All %d are now indexed.", ri, i))
+	return rv
 }
 
 // reindexContractSpecs creates all missing contract specification indexes.
 // This is a function for a migration, not intended for outside use.
-func reindexContractSpecs(ctx sdk.Context, mdKeeper Keeper, lookup *indexLookup) error {
+func reindexContractSpecs(ctx sdk.Context, mdKeeper Keeper, lookup keyLookup) error {
 	store := ctx.KVStore(mdKeeper.storeKey)
-	return mdKeeper.IterateContractSpecs(ctx, func(contractSpec types.ContractSpecification) (stop bool) {
+	i := 0
+	ri := 0
+	rv := mdKeeper.IterateContractSpecs(ctx, func(contractSpec types.ContractSpecification) (stop bool) {
+		i++
+		didri := false
 		for _, key := range getContractSpecIndexValues(&contractSpec).IndexKeys() {
 			if !lookup.has(key) {
 				store.Set(key, []byte{0x01})
+				didri = true
 			}
+		}
+		if didri {
+			ri++
+		}
+		if i%10000 == 0 {
+			ctx.Logger().Info(fmt.Sprintf("Checked %d contract specs and reindexed %d of them.", i, ri))
 		}
 		return false
 	})
+	ctx.Logger().Info(fmt.Sprintf("Done reindexing %d contract specs. All %d are now indexed.", ri, i))
+	return rv
 }
 
-// getEmptySessions finds all sessions that don't have any records.
+// identifySessionsWithRecords identifies all sessions that have records associated with them.
 // This is a function for a migration, not intended for outside use.
-func getEmptySessions(ctx sdk.Context, mdKeeper Keeper) (rv []types.MetadataAddress, err error) {
-	store := ctx.KVStore(mdKeeper.storeKey)
-	sPre := types.SessionKeyPrefix
-	iter := sdk.KVStorePrefixIterator(store, sPre)
+func identifySessionsWithRecords(ctx sdk.Context, mdKeeper Keeper) (rv keyLookup, err error) {
+	rv = newKeyLookup()
+	pre := types.RecordKeyPrefix
+	iter := sdk.KVStorePrefixIterator(ctx.KVStore(mdKeeper.storeKey), pre)
 	defer func() {
 		err = iter.Close()
 	}()
+	i := 0
 	for ; iter.Valid(); iter.Next() {
-		if !mdKeeper.sessionHasRecords(ctx, iter.Key()) {
-			rv = appendMDIfNew(rv, iter.Key())
+		i++
+		var record types.Record
+		merr := mdKeeper.cdc.Unmarshal(iter.Value(), &record)
+		if merr == nil {
+			rv.add(record.SessionId)
+		}
+		if i%10000 == 0 {
+			ctx.Logger().Info(fmt.Sprintf("Checked %d records and found %d session ids so far.", i, len(rv)))
 		}
 	}
+	ctx.Logger().Info(fmt.Sprintf("Done identifying a total of %d session ids used by %d records.", len(rv), i))
+	return
+}
+
+// identifySessionsToDelete identifies all sessions that are not in the provided keyLookup.
+// This is a function for a migration, not intended for outside use.
+func identifySessionsToDelete(ctx sdk.Context, mdKeeper Keeper, sessionsToKeep keyLookup) (rv []types.MetadataAddress, err error) {
+	pre := types.SessionKeyPrefix
+	iter := sdk.KVStorePrefixIterator(ctx.KVStore(mdKeeper.storeKey), pre)
+	defer func() {
+		err = iter.Close()
+	}()
+	i := 0
+	var key []byte
+	for ; iter.Valid(); iter.Next() {
+		i++
+		key = iter.Key()
+		if !sessionsToKeep.has(key) {
+			rv = append(rv, key)
+		}
+		if i%10000 == 0 {
+			ctx.Logger().Info(fmt.Sprintf("Checked %d sessions and found %d without records so far.", i, len(rv)))
+		}
+	}
+	ctx.Logger().Info(fmt.Sprintf("Done identifying a total of %d sessions without records (out of %d sessions).", len(rv), i))
 	return
 }
 
@@ -252,28 +310,12 @@ func getEmptySessions(ctx sdk.Context, mdKeeper Keeper) (rv []types.MetadataAddr
 // This is a function for a migration, not intended for outside use.
 func deleteSessions(ctx sdk.Context, mdKeeper Keeper, sessionsToDelete []types.MetadataAddress) error {
 	store := ctx.KVStore(mdKeeper.storeKey)
-	for _, sessionID := range sessionsToDelete {
+	for i, sessionID := range sessionsToDelete {
 		store.Delete(sessionID)
+		if i%10000 == 9999 {
+			ctx.Logger().Info(fmt.Sprintf("Deleted %d empty sessions, %d to go.", i+1, len(sessionsToDelete)-i-1))
+		}
 	}
+	ctx.Logger().Info(fmt.Sprintf("Done deleting %d empty sessions.", len(sessionsToDelete)))
 	return nil
-}
-
-// appendMDIfNew appends elements to a slice that aren't already in the slice.
-func appendMDIfNew(slice []types.MetadataAddress, elems ...types.MetadataAddress) []types.MetadataAddress {
-	for _, elem := range elems {
-		if !containsMD(slice, elem) {
-			slice = append(slice, elem)
-		}
-	}
-	return slice
-}
-
-// containsMD returns true if the slice contains the provided elem.
-func containsMD(slice []types.MetadataAddress, elem types.MetadataAddress) bool {
-	for _, md := range slice {
-		if elem.Equals(md) {
-			return true
-		}
-	}
-	return false
 }