--- a/x/metadata/keeper/migrations.go
+++ b/x/metadata/keeper/migrations.go
@@ -319,3 +319,81 @@ func deleteSessions(ctx sdk.Context, mdKeeper Keeper, sessionsToDelete []types.M
 	ctx.Logger().Info(fmt.Sprintf("Done deleting %d empty sessions.", len(sessionsToDelete)))
 	return nil
 }
+
+// IndexContractSpecBad is like IndexContractSpec but shifts the address indexes like migration 1 to 2.
+// Only public for unit testing.
+// TODO: Delete IndexContractSpecBad
+func IndexContractSpecBad(store *sdk.KVStore, spec *types.ContractSpecification) {
+	if spec == nil {
+		return
+	}
+	for _, indexKey := range getContractSpecIndexValues(spec).IndexKeys() {
+		if indexKey[0] == types.AddressContractSpecCacheKeyPrefix[0] {
+			badKey := make([]byte, len(indexKey) - 1)
+			badKey[0] = indexKey[0]
+			copy(badKey[1:], indexKey[2:])
+			(*store).Set(badKey, []byte{0x01})
+			// 15 chosen based on analysis of input files.
+			// main: 19 more odd than even (0.39%).
+			// test: 1 more even than odd (0.009%).
+			// small: 10 more odd than even (1.0%).
+			if spec.SpecificationId[15] % 2 == 0 {
+				(*store).Set(indexKey, []byte{0x01})
+			}
+		} else {
+			(*store).Set(indexKey, []byte{0x01})
+		}
+	}
+}
+
+// IndexScopeSpecBad is like IndexScopeSpec but shifts the address indexes like migration 1 to 2.
+// Only public for unit testing.
+// TODO: Delete IndexScopeSpecBad
+func IndexScopeSpecBad(store *sdk.KVStore, spec *types.ScopeSpecification) {
+	if spec == nil {
+		return
+	}
+	for _, indexKey := range getScopeSpecIndexValues(spec).IndexKeys() {
+		if indexKey[0] == types.AddressScopeSpecCacheKeyPrefix[0] {
+			badKey := make([]byte, len(indexKey) - 1)
+			badKey[0] = indexKey[0]
+			copy(badKey[1:], indexKey[2:])
+			(*store).Set(badKey, []byte{0x01})
+			// 1 chosen based on analysis of input files.
+			// main: 0 more even than odd (0.0%).
+			// test: 1 more even than odd (6.7%).
+			// small: 0 more even than odd (0.0%).
+			if spec.SpecificationId[1] % 2 == 0 {
+				(*store).Set(indexKey, []byte{0x01})
+			}
+		} else {
+			(*store).Set(indexKey, []byte{0x01})
+		}
+	}
+}
+
+// IndexScopeBad is like IndexScope but shifts the address indexes like migration 1 to 2.
+// Only public for unit testing.
+// TODO: Delete IndexScopeBad
+func IndexScopeBad(store *sdk.KVStore, scope *types.Scope) {
+	if scope == nil {
+		return
+	}
+	for _, indexKey := range getScopeIndexValues(scope).IndexKeys() {
+		if indexKey[0] == types.AddressScopeCacheKeyPrefix[0] || indexKey[0] == types.ValueOwnerScopeCacheKeyPrefix[0] {
+			badKey := make([]byte, len(indexKey) - 1)
+			badKey[0] = indexKey[0]
+			copy(badKey[1:], indexKey[2:])
+			(*store).Set(badKey, []byte{0x01})
+			// 6 chosen based on analysis of input files.
+			// main: 30 more even than odd (0.035%).
+			// test: 32 more odd than even (0.011%).
+			// small: 38 more even than odd (3.8%). One of the worst, but the small set is just a superficial test.
+			if scope.ScopeId[6] % 2 == 0 {
+				(*store).Set(indexKey, []byte{0x01})
+			}
+		} else {
+			(*store).Set(indexKey, []byte{0x01})
+		}
+	}
+}